src/main/java/com/mangjoo/io/securityexample/security/service/CustomUserDetailsService.java
package com.mangjoo.io.securityexample.security.service;

import com.mangjoo.io.securityexample.application.persistence.MemberEntity;
import com.mangjoo.io.securityexample.application.persistence.MemberJpaRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@RequiredArgsConstructor
@Service
public class CustomUserDetailsService {

    private final MemberJpaRepository memberJpaRepository;

    public MemberEntity loadUserByUsername(String username) {
        return memberJpaRepository.findByUsername(username)
                .orElseThrow(() -> new UsernameNotFoundException("user not found"));
    }
}
